delete confirmation added

+++ src/components/pages/MyToys/MyToys.jsx
@@ -1,6 +1,6 @@
 import { useContext, useEffect, useState } from "react";
-import { RiDeleteBin5Line } from   "react-icons/ri";
-import {GrDocumentUpdate} from "react-icons/gr";
+import { RiDeleteBin5Line } from "react-icons/ri";
+import { GrDocumentUpdate } from "react-icons/gr";
 import { AuthContext } from "../../providers/AuthProvider";
 import Swal from 'sweetalert2';
 
@@ -17,27 +17,41 @@ const MyToysPage = () => {
     }, [user?.email]);
 
     // Function to handle toy deletion
-      const handleDeleteToy = (toyId) => {
-    // Send a request to the server to delete the toy
-        fetch(`https://toy-e-commerece-server.vercel.app/alltoys/${toyId}`, {
-          method: "DELETE",
-        })
-          .then((response) => response.json())
-          .then((data) => {
-            // Remove the deleted toy from the list
-            setToys((prevToys) => prevToys.filter((toy) => toy._id !== toyId));
-            console.log(data);
-            if(data.deletedCount){
-                Swal.fire({
-                    icon: 'error',
-                    title: 'Deleted!',
-                    text: 'Toy successfully deleted!',
-                    confirmButtonText: 'OK'
-                  });
+    const handleDeleteToy = (toyId) => {
+        Swal.fire({
+            title: 'Confirmation',
+            text: 'Are you sure you want to delete this toy?',
+            icon: 'warning',
+            showCancelButton: true,
+            confirmButtonColor: '#d33',
+            cancelButtonColor: '#3085d6',
+            confirmButtonText: 'Delete',
+            cancelButtonText: 'Cancel'
+        }).then((result) => {
+            if (result.isConfirmed) {
+                fetch(`https://toy-e-commerece-server.vercel.app/alltoys/${toyId}`, {
+                    method: 'DELETE',
+                })
+                    .then((response) => response.json())
+                    .then((data) => {
+                        // Remove the deleted toy from the list
+                        setToys((prevToys) => prevToys.filter((toy) => toy._id !== toyId));
+                        console.log(data);
+                        if (data.deletedCount) {
+                            Swal.fire({
+                                icon: 'success',
+                                title: 'Deleted!',
+                                text: 'Toy successfully deleted!',
+                                confirmButtonText: 'OK',
+                            });
+                        }
+                    })
+                    .catch((error) => console.log(error));
             }
-          })
-          .catch((error) => console.log(error));
-      };
+        });
+    };
+
+
 
     return (
         <div className="w-5/6 mx-auto text-center">
@@ -70,7 +84,7 @@ const MyToysPage = () => {
                             <td className="p-2 border">
                                 <button
                                     className="bg-red-500 hover:bg-red-600 text-white px-2 py-2 rounded"
-                                  onClick={() => handleDeleteToy(toy._id)}
+                                    onClick={() => handleDeleteToy(toy._id)}
                                 >
                                     <RiDeleteBin5Line />
                                 </button>
